+++ front-end/src/app/screens/admin/index.tsx
@@ -1,1 +1,1 @@
-export * from "./Admin";
+export * from "./AdminScreen";
